updated CSS styling
