+++ .eslintrc.json
@@ -38,6 +38,7 @@
     "react/jsx-props-no-spreading": 0,
     "react/prop-types": 0,
     "react/jsx-wrap-multilines": 0,
-    "react/jsx-one-expression-per-line": 0
+    "react/jsx-one-expression-per-line": 0,
+    "no-param-reassign": 0
   }
 }
