+++ public/locales/en/translation.yaml
@@ -7,6 +7,7 @@ error:
 # components
 card:
   pipelineDetails: Go to details
+  litmus-e2e: litmuschaos/litmus-e2e
 
 radialChart:
   pass: Pass
