+++ src/components/Loader/styles.js
@@ -1,0 +1,9 @@
+import { makeStyles } from "@material-ui/core";
+
+const useStyles = makeStyles((theme) => ({
+  spinner: {
+    color: theme.palette.primary.main,
+  },
+}));
+
+export default useStyles;
