+++ src/containers/layouts/Center/index.jsx
@@ -1,15 +1,5 @@
 import React from "react";
-import { makeStyles } from "@material-ui/core";
-
-const useStyles = makeStyles({
-  center: {
-    width: "100%",
-    height: "100%",
-    display: "flex",
-    justifyContent: "center",
-    alignItems: "center",
-  },
-});
+import useStyles from "./styles";
 
 const Center = ({ children, className }) => {
   const classes = useStyles();
