+++ src/containers/layouts/Center/styles.js
@@ -1,0 +1,13 @@
+import { makeStyles } from "@material-ui/core";
+
+const useStyles = makeStyles({
+  center: {
+    width: "100%",
+    height: "100%",
+    display: "flex",
+    justifyContent: "center",
+    alignItems: "center",
+  },
+});
+
+export default useStyles;
